requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
cryptography>=41.0.0
anthropic>=0.40.0
reportlab>=4.0.0
Flask-Limiter>=3.8.0
